+++ WCLiquidGlassMenu.h
@@ -39,6 +39,10 @@ void WCLiquidGlassCaptureWCGlassRegistration(NSString *title,
                                              NSString *version,
                                              NSString *controllerName);
 NSDictionary<NSString *, NSString *> * _Nullable WCLiquidGlassCurrentWCGlassRegistration(void);
+void WCLiquidGlassBeginWCGlassPluginListObservation(void);
+void WCLiquidGlassObserveWCGlassPluginListNavigation(UIViewController * _Nullable sourceController,
+                                                     UIViewController * _Nullable destinationController,
+                                                     BOOL completed);
 extern NSString *const WCLiquidGlassManualTextEditNotification;
 
 #ifdef __cplusplus